Stack Machine

The pickle is composed of opcodes and arguments that control over stack machine. A stack machine will be executed with commands and rebuild the same object.

Pickle.OpCodes.OpCode — Type

OpCodes of Pickle stack machine.

Pickle.OpCodes.argument — Method

argument(::OpCode)

return the argument reader of an OpCode.

Pickle.OpCodes.genops — Function

genops(io::IO, yield_end_pos=false)

Generate/Dump all the opcode from pickle io.